% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samplesize.R
\name{calculate_sample_size}
\alias{calculate_sample_size}
\title{Sample size calculation}
\usage{
calculate_sample_size(standard, alph, pw, mde)
}
\arguments{
\item{standard}{The standard deviation}

\item{pw}{The power}

\item{mde}{The minimum delta effect}

\item{alpha}{The alpha}
}
\value{
The sample size
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
sample_size = calculate_sample_size(500,0.05,0.9,200)
}
